sort file structure and relevant imports
