+++ .gitignore
@@ -29,7 +29,7 @@ test_files/qe_output_*.out
 /tests/h2_otf.out-f.xyz-bak
 /tests/pwscf.wfc*
 /tests/pwscf.in
-/tests/pwscf.xml
+/tests/*.xml
 /tests/*-RESTART.kp*
 /tests/*.wfn*
 /tests/*.pwi*
