+++ flare/gp_from_aimd.py
@@ -46,9 +46,10 @@ from flare.output import Output
 from flare.predict import predict_on_atom, predict_on_atom_en, \
     predict_on_structure_par, predict_on_structure_par_en
 from flare.struc import Structure
-from flare.util import element_to_Z, \
+from flare.utils.element_coder import element_to_Z, Z_to_element
+from flare.utils.learner import subset_of_frame_by_element, \
     is_std_in_bound_per_species, is_force_in_bound_per_species, \
-    Z_to_element, subset_of_frame_by_element
+    subset_of_frame_by_element
 from flare.mgp.otf import predict_on_structure_mgp
 from flare.mgp.mgp_en import MappedGaussianProcess
