+++ flare/kernels/kernels.py
@@ -2,7 +2,7 @@ import numpy as np
 from math import exp
 from flare.env import AtomicEnvironment
 from numba import njit
-import flare.cutoffs as cf
+import flare.kernels.cutoffs as cf
 
 # -----------------------------------------------------------------------------
 #                            general helper functions
