+++ flare/kernels/mc_3b_sepcut.py
@@ -1,7 +1,7 @@
 import numpy as np
 from numba import njit
 from math import exp
-import flare.cutoffs as cf
+import flare.kernels.cutoffs as cf
 from flare.kernels.kernels import force_helper, grad_constants, grad_helper, \
     force_energy_helper, three_body_en_helper, three_body_helper_1, \
     three_body_helper_2, three_body_grad_helper_1, three_body_grad_helper_2
