+++ flare/otf.py
@@ -12,7 +12,7 @@ import flare.predict as predict
 from flare import struc, gp, env, md
 from flare.dft_interface import dft_software
 from flare.output import Output
-from flare.util import is_std_in_bound
+from flare.utils.learner import is_std_in_bound
 
 
 class OTF:
