+++ tests/test_flare_io.py
@@ -4,7 +4,7 @@ import numpy as np
 from pymatgen.io.vasp.outputs import Vasprun
 from flare.struc import Structure, get_unique_species
 from flare.dft_interface.vasp_util import md_trajectory_from_vasprun
-from flare.flare_io import md_trajectory_to_file, md_trajectory_from_file
+from flare.utils.flare_io import md_trajectory_to_file, md_trajectory_from_file
 
 pytestmark = pytest.mark.filterwarnings("ignore::UserWarning",\
     "ignore::pymatgen.io.vasp.outputs.UnconvergedVASPWarning")
